% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Benchmarks.R
\name{calcBasicBenchmarks}
\alias{calcBasicBenchmarks}
\title{Calculate Basic Benchmarks for Experimental Data}
\usage{
calcBasicBenchmarks(Stats, StatsPep, Param)
}
\arguments{
\item{Stats}{A data frame containing protein-level statistics from the experimental data.}

\item{StatsPep}{A data frame containing peptide-level statistics from the experimental data.}

\item{Param}{A list of parameters used for the analysis, typically generated by \code{def_param} or \code{def_param_from_yaml}.}
}
\value{
A list containing the following elements:
\describe{
  \item{globalBMs}{A list of global benchmarks calculated at various levels (peptide, protein, PTM). This includes metrics such as the number of peptides and proteins, dynamic range, missing data percentages, and distribution characteristics. The structure of this list includes:}
  \describe{
    \item{numPeptides}{Number of peptides analyzed.}
    \item{numProteins}{Number of proteins analyzed.}
    \item{dynRangePep}{Dynamic range at the peptide level.}
    \item{propUniquePep}{Proportion of unique peptides.}
    \item{uniqueStrippedPep}{Number of unique peptides after stripping modifications.}
    \item{percMissingPep}{Percentage of missing values at the peptide level.}
    \item{propMisCleavedPeps}{Proportion of miscleaved peptides.}
    \item{skewnessPeps}{Skewness of peptide intensities.}
    \item{kurtosisPeps}{Kurtosis of peptide intensities.}
    \item{sdPeps}{Standard deviation of peptide intensities.}
    \item{numQuantProtGroups}{Number of quantified protein groups.}
    \item{dynRangeProt}{Dynamic range at the protein level.}
    \item{propUniqueProts}{Proportion of unique proteins.}
    \item{percMissingProt}{Percentage of missing values at the protein level.}
    \item{meanPepPerProt}{Mean number of peptides per protein.}
    \item{propMisCleavedProts}{Proportion of miscleaved proteins.}
    \item{skewnessProts}{Skewness of protein intensities.}
    \item{kurtosisProts}{Kurtosis of protein intensities.}
    \item{sdProts}{Standard deviation of protein intensities.}
    \item{numProteoforms}{Number of proteoforms identified.}
    \item{numModPeptides}{Number of modified peptides identified.}
    \item{meanProteoformsPerProt}{Mean number of proteoforms per protein.}
    \item{propModAndUnmodPep}{Proportion of modified peptides with an unmodified counterpart.}
    \item{percOverlapModPepProt}{Percentage of overlap between modified peptides and their proteins.}
  }
}
}
\description{
This internal function calculates basic benchmarks for experimental data, focusing on metrics such as the number of peptides and proteins, dynamic range, proportion of unique peptides and proteins, missing data percentages, and measures of distribution such as skewness and kurtosis. This function does not require ground truth data and is useful for summarizing the general characteristics of a dataset.
}
\keyword{internal}
