% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Benchmarks.R
\name{calcROC}
\alias{calcROC}
\title{Calculate ROC Curves and AUC}
\usage{
calcROC(Stats, columnName, groundtruthColumn = "min1Reg")
}
\arguments{
\item{Stats}{A data frame containing statistical results, including the column used for the ROC curve and a ground truth column.}

\item{columnName}{A character string specifying the name of the column in `Stats` used to calculate the ROC curve. This column typically contains the statistical measure of interest (e.g., p-values or adjusted p-values).}

\item{groundtruthColumn}{A character string specifying the name of the column in `Stats` that contains the ground truth labels (e.g., `TRUE` for positives and `FALSE` for negatives). The default is `"min1Reg"`.}
}
\value{
A matrix containing the following columns:
\describe{
  \item{FPR}{False Positive Rate at each threshold.}
  \item{TPR}{True Positive Rate at each threshold.}
  \item{FDR}{The false discovery rate at each threshold.}
  \item{tFDR}{The true false discovery rate at each threshold.}
  \item{AUC}{The calculated Area Under the Curve (AUC). This is a single value representing the area under the ROC curve.}
}
}
\description{
This internal function calculates the Receiver Operating Characteristic (ROC) curve, False Discovery Rate (FDR), and the Area Under the Curve (AUC) based on statistical results. The ROC curve is constructed by varying the threshold of a statistical measure and recording the True Positive Rate (TPR) and False Positive Rate (FPR).
}
\keyword{internal}
