% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GenerateGroundTruth.R
\name{createRegulationPattern}
\alias{createRegulationPattern}
\title{Create a Regulation Pattern}
\usage{
createRegulationPattern(NumCond)
}
\arguments{
\item{NumCond}{An integer specifying the number of conditions.}
}
\value{
A numeric vector of length `NumCond` representing the regulation pattern.
}
\description{
This function generates a regulation pattern for a given number of conditions. The pattern
consists of random values representing up-regulation or down-regulation for each condition.
}
\keyword{internal}
