% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Digestion.R
\name{digestGroundTruth}
\alias{digestGroundTruth}
\title{Perform proteoform digestion on a set of proteoforms with optional parallel computing}
\usage{
digestGroundTruth(proteoforms, parameters, searchIndex = NULL)
}
\arguments{
\item{proteoforms}{A data frame containing the proteoform sequences and associated data.}

\item{parameters}{A list of parameters including enzyme type, number of cores for parallel
computing, maximum number of missed cleavages, peptide length limits, and proportion of missed cleavages.}
}
\value{
A data frame containing the digested peptides, with details on modifications,
mass shifts, and missed cleavages.
}
\description{
This function wraps the \code{proteoformDigestion} function to perform enzymatic digestion
on a set of proteoforms. It supports parallel computing and allows sampling of peptides
based on a distribution derived from \code{PropMissedCleavages} according to their missed
cleavages (MC).
}
\keyword{internal}
