% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{gather_all_sims}
\alias{gather_all_sims}
\title{Gather parameters and benchmarks from all available runs}
\usage{
gather_all_sims(Config, stage = "DataAnalysis")
}
\arguments{
\item{Config}{A list of configuration settings, including file paths and
computational settings, typically generated using \code{set_phosfake}.}
}
\value{
A list containing the parameters and benchmarks from all available runs.
}
\description{
This function retrieves the parameters and benchmarks (if available) from all available runs
and compiles them into a list for further analysis.
}
\examples{
config <- set_phosfake(resultFilePath = tempdir())
all_results <- gather_all_sims(config)

}
