% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{generate_combinations}
\alias{generate_combinations}
\title{Generate all combinations of parameter sets}
\usage{
generate_combinations(params)
}
\arguments{
\item{params}{A list of parameter vectors, where each vector represents 
different possible values for a parameter.}
}
\value{
A list of lists, each sublist containing a unique combination of 
parameters.
}
\description{
This function takes a list of parameter vectors and generates all possible 
combinations. The result is converted into a list of lists, where each 
sublist represents a unique combination of the provided parameters.
}
\examples{
params <- def_param()
params$paramGroundTruth$NumReps <- c(2, 4, 6)
combinations <- generate_combinations(params)
}
