% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{get_bmtitles}
\alias{get_bmtitles}
\title{Function to get human-readable titles for the different parameters}
\usage{
get_bmtitles()
}
\description{
Function to get human-readable titles for the different parameters
}
