% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{matrix_benchmarks}
\alias{matrix_benchmarks}
\title{Create a matrix of benchmarks from simulation results}
\usage{
matrix_benchmarks(allBs, Config)
}
\arguments{
\item{allBs}{A list of results from the \code{run_sims} function, where each
element contains simulation results and benchmarks for a specific parameter
combination.}

\item{Config}{A list of configuration settings, including file paths and
computational settings, typically generated using \code{set_proteomaker}.}
}
\value{
A data frame with rows representing each simulation and columns
representing benchmark values and parameter settings.
}
\description{
This function compiles a matrix of benchmark results from the outputs of multiple
simulations. It extracts benchmark metrics and associated parameter values,
creating a comprehensive data frame for analysis.
}
\examples{
allBs <- list(
  sim1 = list(Benchmarks = list(globalBMs = list(numPeptides = 10)), Param = list(NumReps = 3))
)
benchmark_matrix <- matrix_benchmarks(allBs, list())

}
