% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_MSRun.R
\name{MSRunSim}
\alias{MSRunSim}
\title{Simulate MS run analysis}
\usage{
MSRunSim(Digested, parameters, searchIndex = NULL)
}
\arguments{
\item{Digested}{A data frame containing digested peptidoforms with associated data.}

\item{parameters}{A list containing various parameters for the MS run simulation.}
}
\value{
A data frame similar to `Digested` but with modifications introduced by the MS run simulation.
}
\description{
This function simulates MS run analysis, which includes:
- Adding random noise to each sample due to MS instrument.
- Removing a specific proportion of peptidoforms due to detection limitations.
- Removing a percentage of non-missing intensities based on probability weights depending on intensity values.
- Introducing peptide identification false discovery rate.
- Adding PTMs false localization rate for multiple PTM types.
- Filtering out peptidoforms based on a maximum number of missing values threshold.
}
\keyword{internal}
