% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{param_table}
\alias{param_table}
\title{Create Parameter Table from YAML Configuration}
\usage{
param_table()
}
\value{
A data frame containing the parameters from the YAML file with the following columns:
\describe{
  \item{Category}{The category to which the parameter belongs.}
  \item{Group}{The group within the category.}
  \item{Explanation}{A description of the parameter.}
  \item{MinValue}{The minimum allowable value for the parameter.}
  \item{MaxValue}{The maximum allowable value for the parameter.}
  \item{DefaultValue}{The default value for the parameter.}
}
}
\description{
This function reads a YAML file containing parameter settings, processes them
to handle `NA` values, and converts the parameters into a data frame. The resulting
table includes categories, groups, explanations, and min/max/default values for each parameter.
}
\keyword{internal}
