% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Benchmarks.R
\name{readFlashLFQ}
\alias{readFlashLFQ}
\title{Read and Process FlashLFQ Data}
\usage{
readFlashLFQ(psms, allPeps, Prots, Param = NULL)
}
\arguments{
\item{psms}{A data frame containing PSM-level data from FlashLFQ.}

\item{allPeps}{A data frame containing peptide-level data from FlashLFQ.}

\item{Prots}{A data frame containing protein-level data from FlashLFQ.}

\item{Param}{A list of parameters used for the analysis, including quantification columns and other settings.}
}
\value{
A list containing the processed data:
\describe{
  \item{Param}{The updated parameter list, including quantification column names.}
  \item{allPeps}{The processed peptide-level data.}
  \item{Prots}{The processed protein-level data.}
}
}
\description{
This internal function reads and processes data generated by FlashLFQ, preparing it for downstream analysis. It handles accession numbers, quantification columns, and other relevant information.
}
\keyword{Remove}
\keyword{carba.}
\keyword{flashLFQ}
\keyword{in}
\keyword{internal}
\keyword{output:}
\keyword{the}
