% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Benchmarks.R
\name{readMaxQuant}
\alias{readMaxQuant}
\title{Read and Process MaxQuant Data}
\usage{
readMaxQuant(allPeps, Prots, Param = NULL)
}
\arguments{
\item{allPeps}{A data frame containing peptide-level data from MaxQuant.}

\item{Prots}{A data frame containing protein-level data from MaxQuant.}

\item{Param}{A list of parameters used for the analysis, including quantification columns and other settings.}
}
\value{
A list containing the processed data:
\describe{
  \item{Param}{The updated parameter list, including quantification column names.}
  \item{allPeps}{The processed peptide-level data.}
  \item{Prots}{The processed protein-level data.}
}
}
\description{
This internal function reads and processes data generated by MaxQuant, preparing it for downstream analysis. It extracts relevant information such as accession numbers, missed cleavages, and PTM types, and filters the data based on quantification columns.
}
\keyword{internal}
