% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Benchmarks.R
\name{readProline}
\alias{readProline}
\title{Read and Process Proline Data}
\usage{
readProline(psms, allPeps, Prots, Param = NULL)
}
\arguments{
\item{psms}{A data frame containing PSM-level data from Proline.}

\item{allPeps}{A data frame containing peptide-level data from Proline.}

\item{Prots}{A data frame containing protein-level data from Proline.}

\item{Param}{A list of parameters used for the analysis, including quantification columns and other settings.}
}
\value{
A list containing the processed data:
\describe{
  \item{Param}{The updated parameter list, including quantification column names.}
  \item{allPeps}{The processed peptide-level data.}
  \item{Prots}{The processed protein-level data.}
}
}
\description{
This internal function reads and processes data generated by Proline, preparing it for downstream analysis. It handles missing cleavages, PTM types, and accession numbers, and filters the data based on quantification columns.
}
\keyword{Proline}
\keyword{Remove}
\keyword{carba.}
\keyword{in}
\keyword{internal}
\keyword{output:}
\keyword{the}
