% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Benchmarks.R
\name{readWombat}
\alias{readWombat}
\title{Read and Process Wombat Data}
\usage{
readWombat(allPeps, Prots, Param = NULL)
}
\arguments{
\item{allPeps}{A data frame containing peptide-level data from Wombat.}

\item{Prots}{A data frame containing protein-level data from Wombat.}

\item{Param}{A list of parameters used for the analysis, including quantification columns and other settings.}
}
\value{
A list containing the processed data:
\describe{
  \item{Param}{The updated parameter list, including quantification column names.}
  \item{allPeps}{The processed peptide-level data.}
  \item{Prots}{The processed protein-level data.}
}
}
\description{
This internal function reads and processes data generated by Wombat, preparing it for downstream analysis. It extracts relevant information such as accession numbers, missed cleavages, and PTM types, and filters the data based on quantification columns.
}
\keyword{Proline}
\keyword{Remove}
\keyword{carba.}
\keyword{in}
\keyword{internal}
\keyword{output:}
\keyword{the}
