% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Statistics.R
\name{runPolySTest}
\alias{runPolySTest}
\title{Run PolySTest Statistical Tests on Quantitative Data}
\usage{
runPolySTest(fullData, Param, refCond, onlyLIMMA = F, cores = 1)
}
\arguments{
\item{fullData}{A data frame containing quantitative data, with rows representing features (e.g., proteins or peptides) and columns representing samples. The data frame should also include metadata columns such as `Regulation_Amplitude` and `Regulation_Pattern`.}

\item{Param}{A list of parameters, including:
\describe{
  \item{QuantColnames}{The names of the columns containing the quantitative data to be analyzed.}
  \item{NumCond}{The number of conditions in the experiment.}
  \item{NumReps}{The number of replicates per condition.}
  \item{StatPaired}{A logical value indicating whether the experiment is paired.}
}}

\item{refCond}{A numeric or character value indicating the reference condition to be used in the statistical comparisons.}

\item{onlyLIMMA}{A logical value indicating whether to restrict the analysis to only the `limma` test. Default is `FALSE`, which means all available tests in `PolySTest` are run.}

\item{cores}{An integer specifying the number of cores to use for parallel processing. Default is 1.}
}
\value{
A data frame with the original data and additional columns for log ratios, p-values, FDR values, and ground truth indicators for differential regulation.
}
\description{
This function runs a set of statistical tests on quantitative data using the `PolySTest` package. The function handles paired and unpaired tests, reorganizes the data according to experimental conditions, and returns a comprehensive table with log ratios, p-values, and FDR values for the comparisons. It allows the user to specify the number of cores for parallel computation and to limit the tests to only the `limma` method if desired.
}
\keyword{internal}
