% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GenerateGroundTruth.R
\name{samplePreparation}
\alias{samplePreparation}
\title{Prepare a Sample of Proteoforms}
\usage{
samplePreparation(parameters)
}
\arguments{
\item{parameters}{A list of parameters required for sample preparation, including paths to input files and
modification details.}
}
\value{
A data frame containing all proteoforms generated, including:
\itemize{
  \item Unmodified proteins,
  \item Modified proteoforms,
  \item Modifiable but unmodified proteins.
}
}
\description{
This function generates a set of proteoforms by using the `proteinInput` and `performModification` functions.
It returns a data frame that includes unmodified proteins, modified proteoforms, and modifiable but unmodified
proteins.
}
\details{
The function performs the following steps:
\enumerate{
  \item Imports the protein sequences using `proteinInput`.
  \item Modifies the selected proteins using `performModification`.
  \item Combines unmodified and modified proteoforms into a single data frame.
}
}
\keyword{internal}
