% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{set_proteomaker}
\alias{set_proteomaker}
\title{Set paths and general configuration for ProteoMaker}
\usage{
set_proteomaker(
  fastaFilePath = system.file("Proteomes", package = "ProteoMaker"),
  resultFilePath = "SimulatedDatasets",
  cores = 2,
  clusterType = "PSOCK",
  runStatTests = TRUE,
  calcAllBenchmarks = TRUE
)
}
\arguments{
\item{fastaFilePath}{A character string specifying the path to the directory
containing the FASTA files. Default is `"Proteomes"`.}

\item{resultFilePath}{A character string specifying the path to the directory
where the simulation results will be saved. Default is `"SimulatedDatasets"`.}

\item{cores}{An integer specifying the number of cores to use for parallel
processing. Default is 2.}

\item{clusterType}{A character string specifying the type of cluster to use
for parallel computing (e.g., `"FORK"`, `"PSOCK"`). Default is `"FORK"`.}

\item{runStatTests}{A logical value indicating whether to run statistical tests. Default is `TRUE`.}

\item{calcAllBenchmarks}{A logical value indicating whether to calculate all
benchmarks. Default is `TRUE`.}
}
\value{
A list containing the configuration settings:
\describe{
  \item{fastaFilePath}{The path to the FASTA files.}
  \item{resultFilePath}{The path to the result files.}
  \item{cores}{The number of cores used.}
  \item{clusterType}{The type of cluster used.}
  \item{calcAllBenchmarks}{Whether to calculate all benchmarks.}
  \item{runStatTests}{Whether to run statistical tests.}
}
}
\description{
This function sets the paths and general configuration parameters needed for
running simulations with the ProteoMaker package. It creates the necessary output
directories and returns a list of configuration settings.
}
\examples{
config <- set_proteomaker()
config <- set_proteomaker(fastaFilePath = "CustomProteomes",
resultFilePath = "Results", cores = 4, clusterType = "PSOCK")
}
