% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{gather_all_sims}
\alias{gather_all_sims}
\title{Gather parameters and benchmarks from all available runs}
\usage{
gather_all_sims(Config, stage = "DataAnalysis")
}
\arguments{
\item{Config}{A list of configuration settings, including file paths and
computational settings, typically generated using \code{set_proteomaker}.}

\item{stage}{A character string indicating the stage of the simulation for
which the output is requested. Default is \code{"DataAnalysis"}. Values are \code{"GroundTruth"},
\code{"ProteoformAb"}, \code{"Digest"}, \code{"MSRun"}, and \code{"DataAnalysis"}. Default is \code{"DataAnalysis"}.}
}
\value{
A list containing the parameters and benchmarks from all available runs.
}
\description{
This function retrieves the parameters and benchmarks (if available) from all available runs
and compiles them into a list for further analysis.
}
\examples{
config <- set_proteomaker(resultFilePath = tempdir())
all_results <- gather_all_sims(config)

}
