% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{get_simulation}
\alias{get_simulation}
\title{Retrieve intermediate outputs for a simulation}
\usage{
get_simulation(Param, Config, stage = "DataAnalysis")
}
\arguments{
\item{Param}{A list of parameters used in the simulation, typically a subset
of those generated by functions like \code{def_param}.}

\item{Config}{A list of configuration settings, including file paths and
computational settings, typically generated using \code{set_phosfake}.}

\item{stage}{A character string indicating the stage of the simulation for
which the output is requested. Default is \code{"DataAnalysis"}. Values are \code{"GroundTruth"},
\code{"ProteoformAb"}, \code{"Digest"}, \code{"MSRun"}, and \code{"DataAnalysis"}.}
}
\value{
A list containing the loaded data objects, including \code{Stats},
\code{StatsPep}, and \code{Benchmarks}, if they exist. If no file is found,
the function returns \code{NULL}.
}
\description{
This function retrieves intermediate outputs for a simulation based on the
provided parameters and configuration. It checks for the existence of the
output file corresponding to a specific simulation stage and loads the
data if available.
}
\examples{
config <- set_phosfake(resultFilePath = tempdir())
Param <- def_param()
Param$paramGroundTruth$NumReps <- 5
benchmarks <- run_sims(Param, config)
result <- get_simulation(benchmarks[[1]]$Param, config, stage="MSRun")
}
