% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GenerateGroundTruth.R
\name{modify_seq}
\alias{modify_seq}
\title{Modify single sequence}
\usage{
modify_seq(seq, pars)
}
\arguments{
\item{seq}{A character vector of protein sequences to be modified.}

\item{pars}{A list of parameters for the modification process, including:
 \itemize{
 \item \code{ModifiableResidues}: A list of amino acid residues that can be modified for each PTM type.
 \item \code{PTMTypes}: A character vector of modification types.
 \item \code{PTMTypesDistr}: A numeric vector representing the background frequencies for each PTM type.
 \item \code{ModifiableResiduesDistr}: A list of numeric vectors representing the background frequencies
 for each modifiable residue within each PTM type.
 \item \code{PTMMultipleLambda}: A numeric value specifying the lambda parameter for the truncated Poisson distribution,
 which determines the number of modifications per sequence.
 }


 keywords internal}
}
\description{
This function modified based on the modification type and residue frequencies.
}
