% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DataAnalysis.R
\name{proteinSummarisation}
\alias{proteinSummarisation}
\title{Summarize Protein Abundance from Peptide-Level Data}
\usage{
proteinSummarisation(peptable, parameters)
}
\arguments{
\item{peptable}{A data frame containing peptide-level data, including accession numbers, sequence, PTM types, and quantification columns.}

\item{parameters}{A list of parameters, including:
\describe{
  \item{ProtSummarization}{The method used for protein summarization, e.g., "sum.top3" or "medpolish".}
  \item{MinUniquePep}{The minimum number of unique peptides required to summarize a protein.}
  \item{QuantColnames}{The names of the columns containing quantification data.}
  \item{Cores}{The number of cores to use for parallel processing.}
  \item{ClusterType}{The type of cluster to use for parallel processing (e.g., "FORK", "PSOCK").}
}}
}
\value{
A data frame containing summarized protein-level data, where each row represents a protein, and the columns include protein information and summarized quantification data.
}
\description{
This function summarizes protein abundance from peptide-level data. It removes modified peptides, groups peptides by their parent proteins, and applies a summarization method (e.g., sum of top 3, median polish) to estimate protein abundance. The function supports parallel processing to speed up the summarization process.
}
\keyword{internal}
