% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{render_benchmark_table}
\alias{render_benchmark_table}
\title{Render a Benchmark Summary Table}
\usage{
render_benchmark_table(benchmatrix, benchmarks = NULL)
}
\arguments{
\item{benchmatrix}{A data frame or matrix containing benchmark metrics.}

\item{benchmarks}{A character vector of benchmark variable names, or a numeric vector of column indices.
If NULL, the first 12 defined benchmarks will be used.}
}
\value{
A data frame summarizing the selected benchmark metrics.
}
\description{
This function summarizes benchmark results from a matrix of metrics
(typically produced by proteomics simulations or evaluation pipelines).
It returns a human-readable table with either raw values (if one row)
or value ranges (if multiple rows). It also prints a markdown version
of the table in the console
}
