% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{run_sims}
\alias{run_sims}
\title{Run all simulations for given parameters and configurations}
\usage{
run_sims(Parameters, Config)
}
\arguments{
\item{Parameters}{A list containing categorized parameter sets, typically
generated using \code{def_param}.}

\item{Config}{A list containing configuration settings, such as file paths
and computational settings, typically generated using \code{set_phosfake}.}
}
\value{
A list of results, where each element contains simulation results
and benchmarks for a specific parameter combination.
}
\description{
This function runs simulations based on the provided parameter sets and
configuration settings. It runs the given parameter settings, processes
ground truth data, simulates proteoform abundance, performs digestion,
conducts MS runs, and carries out data analysis. Results and benchmarks are
saved to the output directory specified in \code{def_param}.
}
\examples{
params <- def_param()
config <- set_phosfake()
results <- run_sims(params, config)
}
