% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_BatchRunFuncs.R
\name{visualize_benchmarks}
\alias{visualize_benchmarks}
\title{Visualize benchmarks vs 1 or two parameters}
\usage{
visualize_benchmarks(
  benchmatrix,
  benchmarks = NULL,
  ref_par = "WrongIDs",
  fullrange = FALSE,
  cols = NULL,
  errorbar = FALSE
)
}
\arguments{
\item{benchmatrix}{A data frame containing benchmark results, typically
generated by the \code{matrix_benchmarks} function.}

\item{benchmarks}{A character vector specifying the names of the benchmarks
to be visualized. If \code{NULL}, all benchmarks will be plotted.}

\item{ref_par}{A character vector specifying the names of the parameters to
be used for visualization. It can contain one or two parameters.}

\item{fullrange}{A logical value indicating whether to use the maximal range of
set of benchmarks (works only for one `ref_par`)}

\item{cols}{A character vector specifying the colors to be used for the
visualization. If \code{NULL}, default colors will be used.}

\item{errorbar}{A logical value indicating whether to show mean +/- standard
deviations in the
visualization. Default is \code{FALSE}. This parameter is not applied to the color maps.}
}
\value{
A bar plot or a colored map visualizing the specified benchmarks against the
reference parameters.
}
\description{
This function visualizes the benchmarks from a matrix of benchmark results. It creates
bar plots for the specified benchmarks and parameters, allowing for the
comparison of benchmark values across different parameter settings.
}
\examples{
conf <- set_proteomaker(resultFilePath = tempdir())
results <- run_sims(def_param(), conf)
benchmark_matrix <- matrix_benchmarks(results, conf)
visualize_benchmarks(benchmark_matrix, ref_par = "NumReps")

}
